Added screen size constants. Fixed warnings related to implicit casting.
Spent a few hours trying to implement NDC. It failed - I broke everything. Will try again after I've mentally recovered.

+++ Watermelon/src/framework/OrthographicCamera.cpp
@@ -17,7 +17,7 @@ void OrthographicCamera::SetFocus(const GameObject* focusPtr) {
 }
 
 Vec2 OrthographicCamera::GetPosition() const {
-	return Vec2(cam->x, cam->y);
+	return Vec2((float)cam->x, (float)cam->y);
 }
 
 SDL_Rect* OrthographicCamera::GetRect() const {
